Initial library (#1)
This contains the initial creation of this library. It contains the forked, cleaned up, and namespaced copy of resources-poet in addition to the api, core, and top-level modules of the stylist library. A sample providers module and library module are also included, but will be used in a sample app in a followup.